+++ Blurred-ios/Channels/Video/ChannelVideoCellNode.swift
@@ -31,17 +31,16 @@ class ChannelVideoCellNode: ASCellNode {
         }
         self.addSubnode(self.videoNode)
         self.addSubnode(self.gradientNode)
-        
         DispatchQueue.main.async() {
             let overlay = ChannelVideoOverlayView()
             overlay.frame = CGRect(x: 0, y: 0, width: UIScreen.main.bounds.size.width, height: UIScreen.main.bounds.size.height)
             self.view.addSubview(overlay)
-            overlay.videoId = videoId
-            overlay.sendRequest()
-            overlay.sendCheckLikeRequest()
-            overlay.checkLikeCount()
         }
-        
+        let touches = ChannelVideoOverlayTouch()
+        touches.videoId = videoId
+        touches.sendRequest()
+        touches.sendCheckLikeRequest()
+        touches.checkLikeCount()
   }
   override func layoutSpecThatFits(_ constrainedSize: ASSizeRange) -> ASLayoutSpec {
     let ratio = UIScreen.main.bounds.height / UIScreen.main.bounds.width